Simplify mypy invocation
