+++ check.sh
@@ -16,11 +16,7 @@ fi
 
 black --check .
 ruff check .
-mypy --ignore-missing-imports --explicit-package-bases --no-site-packages main.py shared tests || true
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F1
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F2
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F3
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F4
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F5
-mypy --ignore-missing-imports --strict --explicit-package-bases --no-site-packages --exclude '(acceptance_tests|unit_tests)' features/F6
-pytest -q tests features/*/unit_tests
+mypy --ignore-missing-imports --strict --explicit-package-bases \
+  --no-site-packages --exclude '(acceptance_tests|unit_tests)' \
+  main.py shared features
+pytest -q features/*/unit_tests
